fix: build_device_config skips duplicate jobs
Fixing this issue also caused circular import issues, so the code in
'misc' and some code in 'utils' were moved into other locations to
mitigate the occurrence of circular imports.

Also updated set_expired calls with the correct arguments.

ZEN-34698
ZEN-34735
