+++ Products/ZenCollector/configcache/task.py
@@ -20,7 +20,8 @@ from Products.ZenUtils.RedisUtils import getRedisClient, getRedisUrl
 from Products.Jobber.task import requires, DMD, Abortable
 from Products.Jobber.zenjobs import app
 
-from .cache import ConfigKey, ConfigQuery, ConfigRecord
+from .cache import ConfigKey, ConfigQuery, ConfigRecord, ConfigStatus
+from .propertymap import DevicePropertyMap
 
 
 @app.task(
@@ -47,6 +48,23 @@ def build_device_config(self, monitorname, deviceid, configclassname):
     svcconfigclass = resolve(configclassname)
     svcname = configclassname.rsplit(".", 1)[0]
     store = _getStore()
+    key = ConfigKey(svcname, monitorname, deviceid)
+
+    # Check whether this is an old job, i.e. job pending timeout.
+    # If it is an old job, skip it, manager already sent another one.
+    statuses = tuple(store.get_status(key))
+    if statuses:
+        key, status = statuses[0]
+        if isinstance(status, ConfigStatus.Pending):
+            pendinglimitmap = DevicePropertyMap.make_pending_timeout_map(
+                self.dmd.Devices
+            )
+            now = time()
+            uid = store.get_uid(key.device)
+            duration = pendinglimitmap.get(uid)
+            if status.submitted < (now - duration):
+                return
+
     # Change the configuration's status from 'pending' to 'building' so
     # that configcache-manager doesn't prematurely timeout the build.
     store.set_building((ConfigKey(svcname, monitorname, deviceid), time()))
@@ -99,6 +117,59 @@ def build_device_config(self, monitorname, deviceid, configclassname):
             monitorname,
             svcname,
         )
+
+
+class BuildConfigTaskDispatcher(object):
+    """Encapsulates the act of dispatching the build_device_config task."""
+
+    def __init__(self, configClasses):
+        """
+        Initialize a BuildConfigTaskDispatcher instance.
+
+        The `configClasses` parameter should be the classes used to create
+        the device configurations.
+
+        @type configClasses: Sequence[Class]
+        """
+        self._classnames = {
+            cls.__module__: ".".join((cls.__module__, cls.__name__))
+            for cls in configClasses
+        }
+
+    def dispatch_all(self, monitorid, deviceid, timeout):
+        """
+        Submit a task to build a device configuration from each
+        configuration service.
+        """
+        soft_limit, hard_limit = _get_limits(timeout)
+        for name in self._classnames.values():
+            build_device_config.apply_async(
+                args=(monitorid, deviceid, name),
+                soft_time_limit=soft_limit,
+                time_limit=hard_limit,
+            )
+
+    def dispatch(self, servicename, monitorid, deviceid, timeout):
+        """
+        Submit a task to build device configurations for the specified device.
+
+        @type servicename: str
+        @type monitorid: str
+        @type deviceId: str
+        """
+        name = self._classnames.get(servicename)
+        if name is None:
+            raise ValueError("service name '%s' not found" % servicename)
+        soft_limit, hard_limit = _get_limits(timeout)
+        build_device_config.apply_async(
+            args=(monitorid, deviceid, name),
+            soft_time_limit=soft_limit,
+            time_limit=hard_limit,
+        )
+
+
+def _get_limits(timeout):
+    return timeout, (timeout + (timeout * 0.1))
 
 
 def _getStore():
