+++ Products/ZenCollector/configcache/utils/__init__.py
@@ -9,17 +9,11 @@
 
 from __future__ import absolute_import
 
-from .constants import Constants
-from .dispatcher import BuildConfigTaskDispatcher
 from .pollers import RelStorageInvalidationPoller
-from .propertymap import DevicePropertyMap
 from .services import getConfigServices
 
 
 __all__ = (
-    "BuildConfigTaskDispatcher",
-    "Constants",
-    "DevicePropertyMap",
     "getConfigServices",
     "RelStorageInvalidationPoller",
 )
